Add a first abstraction layer for SegmentRegisterIndex.
When accessing a memory reference that includes a segment register,
give ourselves a way to remember the segment register being used
instead of just remembering its value.

MemoryOrRegisterReference now remembers a SegmentRegisterIndex
instead of a WORD for the segment. The CPU gains read/writeMemory
helpers that take a SegmentRegisterIndex.

This will come in handy as we improve the selector implementation.