Use the CPU's SegmentSelectors in real mode as well.
This isn't quite where I want it yet, but now the SegmentSelector
has a realMode flag that tells the CPU to just access memory relative
to its base (which is computed using the standard "segment << 16".)

This makes it possible to access memory through a segment selector
that was loaded in real mode, after entering protected mode.

Also some work on various protected mode insns.