Get rid of CS:0 "code cache" pointer.
This fixes the last issue with CWSDPMI where we'd loaded everything
and wanted to enter the client program at 00a7:00001100, but because
CS changed, we'd try to get a memoryPointer() for 00a7:00000000 and
this caused a #PF.

For now we fall back to using readMemory() when fetching instructions.
This is a serious slowdown, so I have to fix it, but let's land this
for correctness.

CWSDPMI can now run the i386nf.exe emulation correctness test, woot!