Pass along the original JumpType when calling through a gate.
This fixes a bug where we'd forget to switch the stacks.
Also pass along the Gate itself when jumping through a gate
(both for interrupt/trap gates and call gates) and always
pick up the bit size from the gate when one is present.

This fixes a bunch of racy #GP's in CWSDPMI.